+++ CHANGES.md
@@ -32,6 +32,7 @@
 * We now need `Sequence::Row` instances in `Sequence` to adhere to the Adds specification.
 * Move `Sequence::IndexError` to `Activity::Adds::IndexError` in the `trailblazer-activity` gem. Remove `IndexError#step_id`.
 * Rename `Linear::State::Normalizer` to `Linear::Normalizer::Normalizers` as it represents a container for normalizers.
+* Rename `Linear::State` to `Linear::Sequencer`. This is now a stateless function, only.
 
 # 0.5.0